pyproject.toml
[tool.poetry]
name = "fastapi_jwt"
version = "0.1.0"
description = "Example project to integrate JWT authentication into FastAPI."
authors = ["Johannes Gontrum <j@gontrum.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.36.0"
requests = "^2.22"
python-jose = "^3.0"
python-dotenv = "^0.10.3"
uvicorn = "^0.8.6"

[tool.poetry.dev-dependencies]
black = {version = "^18.3-alpha.4",allows-prereleases = true}
pre-commit = "^1.17"
flake8 = "^3.7"

[tool.black]
line-length = 90
target_version = ["py37"]
exclude = "\\.git"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
